Only check for interference edges with remaining variables
This is probably known, but I discovered it myself, and I think it's kinda neat. Since the edges are undirected, you don't need to check both permutations of an edge (eg. (i, j) and (j, i); both are equivalent). In this case, i and j correspond to the outer and inner loop indices, respectively. Crucially, we know that when we're on iteration i, as long as i is always in ascending order, we know that we've already checked for edges with i in the second position in previous iterations of i. Thus, we only need to check for edges with i in the first position and j in positions > i. This cuts the search space in half (we traverse a right triangle rather than a square).