Greedy scheduling
Now we construct a proper DDG from the program representation, and greedily schedule that with respect to its dependencies (greedy topological sort). The idea here is that even a greedy algorithm should lead to a schedule with sufficiently low register pressure, unlike what I had before, which ended up scheduling most of the loads first in most of the test cases, which would have eventually blown up as soon as we started matrix/vector muls.