Properly handle dead loads
When thinking about the previous setup, I realized there was a subtle bug that could occur, and sure enough, I was able to construct a pathological test that triggered it.

Essentially what would happen is this: a load always loads into a variable. However, it's possible that a load is created in the original schedule whose output variable is never used (a "dead load"). In that case, liveness analysis wouldn't create any interferences for the output variable, because its "transient liveness" wouldn't be detected. Register allocation, upon seeing a variable that apparently doesn't interfere with any other variables, would then have the freedom to allocate any register it wanted, and would just pick one. Unfortunately, it's entirely possible for that register to have been assigned to another variable whose valid liveness range covered the load instruction, which would mean that when the load occurred at runtime, its value would be clobbered.

One strategy to fix this would be to adjust liveness analysis to gracefully handle such "transient" variable lifetimes. This probably wouldn't be terribly difficult, and it could be useful if similar cases were to arise in the future.

However, there's another, more attractive solution. The only reason why we still wanted to execute dead loads in the first place is so that the hardware context's input stream offset could be incremented such that it would have the expected value for subsequent loads from the same stream. This works, but there's an arguably simpler way to handle it in hardware: instead of expecting the offset to be incremented on each stream access, we could keep its value fixed for the duration of a thread's execution, and encode an explicit offset for each load/store instruction, that would be added to the stream offset to form the final stream offset for the load/store.

This solves the problem by allowing us to drop dead load instructions from the schedule entirely (just like any other instructions not reachable during DDG construction), instead of always executing them and ignoring the data, which also means we save some cycles in these cases (both in the core, but also in potential cache misses that may have occurred). Even better, it means we get to remove the additional DDG edges between loads/stores to the same streams, since we don't actually have to guarantee their order anymore. Incidentally, this solves _another_ class of pathological cases as well: when scheduling ends up trying to schedule the _last_ load to a given input stream _first_, all other loads to that stream would have been scheduled before it, thus ballooning register pressure, which is the thing we're trying to avoid by having a DDG/scheduler in the first place!