Initial Uniform mem/instruction impl
Capacity isn't decided, except indirectly by the size of the offset field currently. I'm also thinking I want to add a register offset in addition to the hardcoded one (for LUTs etc), which will likely mean that we'll also want a way to generate (simple) constant values for regs and/or an always-zero reg. The program/compiler components will also need knowledge of this, with perhaps a nice way to alloc uniform space for the former, and passing on symbol information for the latter.